make epsilon and momentum double passby value

+++ projects/hipdnn/sdk/tests/test_utilities/TestCpuFpReferenceBatchnorm.cpp
@@ -658,12 +658,7 @@ TEST(TestCpuFpReferenceBatchnormFp32, BatchnormFwdTrainingNchwBasic)
 
     // Call without optional tensors
     CpuFpReferenceBatchnormImpl<float, float>::batchnormFwdTraining(
-        inputTensor,
-        scaleTensor,
-        biasTensor,
-        outputTensor,
-        static_cast<float>(BATCHNORM_DEFAULT_EPSILON),
-        0.1f);
+        inputTensor, scaleTensor, biasTensor, outputTensor, BATCHNORM_DEFAULT_EPSILON, 0.1);
 }
 
 TEST(TestCpuFpReferenceBatchnormFp32, BatchnormFwdTrainingNchwWithSavedStats)
@@ -683,15 +678,14 @@ TEST(TestCpuFpReferenceBatchnormFp32, BatchnormFwdTrainingNchwWithSavedStats)
     }
 
     // Call with saved statistics for backward pass
-    CpuFpReferenceBatchnormImpl<float, float>::batchnormFwdTraining(
-        inputTensor,
-        scaleTensor,
-        biasTensor,
-        outputTensor,
-        static_cast<float>(BATCHNORM_DEFAULT_EPSILON),
-        0.1f,
-        &savedMean,
-        &savedInvVariance);
+    CpuFpReferenceBatchnormImpl<float, float>::batchnormFwdTraining(inputTensor,
+                                                                    scaleTensor,
+                                                                    biasTensor,
+                                                                    outputTensor,
+                                                                    BATCHNORM_DEFAULT_EPSILON,
+                                                                    0.1,
+                                                                    &savedMean,
+                                                                    &savedInvVariance);
 
     // Verify saved statistics were populated
     for(int i = 0; i < 3; i++)
@@ -721,22 +715,21 @@ TEST(TestCpuFpReferenceBatchnormFp32, BatchnormFwdTrainingNchwWithRunningStats)
         prevRunningVariance.setHostValue(1.0f, 0, i);
     }
 
-    float momentum = 0.1f;
+    double momentum = 0.1;
 
     // Call with running statistics
-    CpuFpReferenceBatchnormImpl<float, float>::batchnormFwdTraining(
-        inputTensor,
-        scaleTensor,
-        biasTensor,
-        outputTensor,
-        static_cast<float>(BATCHNORM_DEFAULT_EPSILON),
-        momentum,
-        nullptr,
-        nullptr,
-        &prevRunningMean,
-        &prevRunningVariance,
-        &nextRunningMean,
-        &nextRunningVariance);
+    CpuFpReferenceBatchnormImpl<float, float>::batchnormFwdTraining(inputTensor,
+                                                                    scaleTensor,
+                                                                    biasTensor,
+                                                                    outputTensor,
+                                                                    BATCHNORM_DEFAULT_EPSILON,
+                                                                    momentum,
+                                                                    nullptr,
+                                                                    nullptr,
+                                                                    &prevRunningMean,
+                                                                    &prevRunningVariance,
+                                                                    &nextRunningMean,
+                                                                    &nextRunningVariance);
 
     // Verify running statistics were updated
     for(int i = 0; i < 3; i++)
@@ -786,19 +779,18 @@ TEST(TestCpuFpReferenceBatchnormFp32, BatchnormFwdTrainingNchwFullFeatures)
     }
 
     // Call with all optional parameters
-    CpuFpReferenceBatchnormImpl<float, float>::batchnormFwdTraining(
-        inputTensor,
-        scaleTensor,
-        biasTensor,
-        outputTensor,
-        static_cast<float>(BATCHNORM_DEFAULT_EPSILON),
-        0.1f,
-        &savedMean,
-        &savedInvVariance,
-        &prevRunningMean,
-        &prevRunningVariance,
-        &nextRunningMean,
-        &nextRunningVariance);
+    CpuFpReferenceBatchnormImpl<float, float>::batchnormFwdTraining(inputTensor,
+                                                                    scaleTensor,
+                                                                    biasTensor,
+                                                                    outputTensor,
+                                                                    BATCHNORM_DEFAULT_EPSILON,
+                                                                    0.1,
+                                                                    &savedMean,
+                                                                    &savedInvVariance,
+                                                                    &prevRunningMean,
+                                                                    &prevRunningVariance,
+                                                                    &nextRunningMean,
+                                                                    &nextRunningVariance);
 
     // Verify all outputs were populated
     for(int i = 0; i < 3; i++)
@@ -844,8 +836,8 @@ TEST(TestCpuFpReferenceBatchnormBfp16, BatchnormFwdTrainingNchw)
         scaleTensor,
         biasTensor,
         outputTensor,
-        static_cast<float>(BATCHNORM_DEFAULT_EPSILON),
-        0.1f,
+        BATCHNORM_DEFAULT_EPSILON,
+        0.1,
         &savedMean,
         &savedInvVariance);
 }
@@ -866,15 +858,14 @@ TEST(TestCpuFpReferenceBatchnormFp16, BatchnormFwdTrainingNchw)
         biasTensor.setHostValue(0.0f, 0, i);
     }
 
-    CpuFpReferenceBatchnormImpl<half, float>::batchnormFwdTraining(
-        inputTensor,
-        scaleTensor,
-        biasTensor,
-        outputTensor,
-        static_cast<float>(BATCHNORM_DEFAULT_EPSILON),
-        0.1f,
-        &savedMean,
-        &savedInvVariance);
+    CpuFpReferenceBatchnormImpl<half, float>::batchnormFwdTraining(inputTensor,
+                                                                   scaleTensor,
+                                                                   biasTensor,
+                                                                   outputTensor,
+                                                                   BATCHNORM_DEFAULT_EPSILON,
+                                                                   0.1,
+                                                                   &savedMean,
+                                                                   &savedInvVariance);
 }
 
 TEST(TestCpuFpReferenceBatchnormFp64, BatchnormFwdTrainingNchw)
@@ -992,12 +983,7 @@ TEST(TestCpuFpReferenceBatchnormFp32, BatchnormFwdTraining2D)
     }
 
     CpuFpReferenceBatchnormImpl<float, float>::batchnormFwdTraining(
-        inputTensor,
-        scaleTensor,
-        biasTensor,
-        outputTensor,
-        static_cast<float>(BATCHNORM_DEFAULT_EPSILON),
-        0.1f);
+        inputTensor, scaleTensor, biasTensor, outputTensor, BATCHNORM_DEFAULT_EPSILON, 0.1);
 }
 
 TEST(TestCpuFpReferenceBatchnormFp32, BatchnormFwdTraining3D)
@@ -1016,12 +1002,7 @@ TEST(TestCpuFpReferenceBatchnormFp32, BatchnormFwdTraining3D)
     }
 
     CpuFpReferenceBatchnormImpl<float, float>::batchnormFwdTraining(
-        inputTensor,
-        scaleTensor,
-        biasTensor,
-        outputTensor,
-        static_cast<float>(BATCHNORM_DEFAULT_EPSILON),
-        0.1f);
+        inputTensor, scaleTensor, biasTensor, outputTensor, BATCHNORM_DEFAULT_EPSILON, 0.1);
 }
 
 TEST(TestCpuFpReferenceBatchnormFp32, BatchnormFwdTrainingNcdhw)
@@ -1041,15 +1022,14 @@ TEST(TestCpuFpReferenceBatchnormFp32, BatchnormFwdTrainingNcdhw)
         biasTensor.setHostValue(0.0f, 0, i);
     }
 
-    CpuFpReferenceBatchnormImpl<float, float>::batchnormFwdTraining(
-        inputTensor,
-        scaleTensor,
-        biasTensor,
-        outputTensor,
-        static_cast<float>(BATCHNORM_DEFAULT_EPSILON),
-        0.1f,
-        &savedMean,
-        &savedInvVariance);
+    CpuFpReferenceBatchnormImpl<float, float>::batchnormFwdTraining(inputTensor,
+                                                                    scaleTensor,
+                                                                    biasTensor,
+                                                                    outputTensor,
+                                                                    BATCHNORM_DEFAULT_EPSILON,
+                                                                    0.1,
+                                                                    &savedMean,
+                                                                    &savedInvVariance);
 }
 
 TEST(TestCpuFpReferenceBatchnormFp32, BatchnormFwdTrainingNdhwc)
@@ -1069,13 +1049,12 @@ TEST(TestCpuFpReferenceBatchnormFp32, BatchnormFwdTrainingNdhwc)
         biasTensor.setHostValue(0.0f, 0, i);
     }
 
-    CpuFpReferenceBatchnormImpl<float, float>::batchnormFwdTraining(
-        inputTensor,
-        scaleTensor,
-        biasTensor,
-        outputTensor,
-        static_cast<float>(BATCHNORM_DEFAULT_EPSILON),
-        0.1f,
-        &savedMean,
-        &savedInvVariance);
+    CpuFpReferenceBatchnormImpl<float, float>::batchnormFwdTraining(inputTensor,
+                                                                    scaleTensor,
+                                                                    biasTensor,
+                                                                    outputTensor,
+                                                                    BATCHNORM_DEFAULT_EPSILON,
+                                                                    0.1,
+                                                                    &savedMean,
+                                                                    &savedInvVariance);
 }
